requirements.txt
# Web Framework & Server
Flask==3.0.0
gunicorn==21.2.0

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

# Environment & Logic
python-dotenv==1.0.0

# Data & Analytics (needed for your CSV/Excel files)
pandas==2.1.3
openpyxl==3.1.2